+++ Backend/MessageService/include/messageservice/controller.h
@@ -24,7 +24,7 @@ class Controller : public IController {
   Controller(IRabitMQClient* mq_client,
               MessageManager* manager, IThreadPool* pool, IConfigProvider* provider = &ProdConfigProvider::instance());
   std::vector<Message> getMessages(const GetMessagePack&) override;
-  std::vector<MessageStatus> getMessagesStatus(const std::vector<Message>&, int receiver_id) override;
+  std::vector<MessageStatus> getMessagesStatus(const std::vector<Message>&, long long receiver_id) override;
 
  protected:
   void subscribeToSaveMessage();
